+++ CHANGELOG.md
@@ -7,3 +7,5 @@
 ### Added
 
 * welcome screen
+* barebone connector view triggered by the Coder's welcome view.
+* back button to return to the main welcome view